+++ src/main/kotlin/com/ampnet/reportservice/controller/ReportingController.kt
@@ -5,9 +5,7 @@ import com.ampnet.reportservice.controller.pojo.TransactionServiceRequest
 import com.ampnet.reportservice.service.ReportingService
 import mu.KLogging
 import org.springframework.format.annotation.DateTimeFormat
-import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -30,7 +28,7 @@ class ReportingController(
         logger.debug { "Received request to get report of transactions for user with uuid: ${userPrincipal.uuid}" }
         val periodRequest = PeriodServiceRequest(from, to)
         val pdfContents = reportingService.generatePdfReportForUserTransactions(userPrincipal.uuid, periodRequest)
-        return ResponseEntity(pdfContents, getHttpHeadersForPdf(), HttpStatus.OK)
+        return ResponseEntity(pdfContents, ControllerUtils.getHttpHeadersForPdf(), HttpStatus.OK)
     }
 
     @GetMapping("/report/user/transaction")
@@ -46,26 +44,6 @@ class ReportingController(
         }
         val transactionServiceRequest = TransactionServiceRequest(userPrincipal.uuid, txHash, fromTxHash, toTxHash)
         val pdfContents = reportingService.generatePdfReportForUserTransaction(transactionServiceRequest)
-        return ResponseEntity(pdfContents, getHttpHeadersForPdf(), HttpStatus.OK)
-    }
-
-    @GetMapping("/report/admin/user")
-    fun getActiveUsersReport(
-        @RequestParam(name = "from", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) from: LocalDate?,
-        @RequestParam(name = "to", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) to: LocalDate?
-    ): ResponseEntity<ByteArray> {
-        val userPrincipal = ControllerUtils.getUserPrincipalFromSecurityContext()
-        logger.debug {
-            "Received request to get users accounts summary for all the active users"
-        }
-        val periodRequest = PeriodServiceRequest(from, to)
-        val pdfContents = reportingService.generatePdfReportForAllActiveUsers(userPrincipal, periodRequest)
-        return ResponseEntity(pdfContents, getHttpHeadersForPdf(), HttpStatus.OK)
-    }
-
-    private fun getHttpHeadersForPdf(): HttpHeaders {
-        val httpHeaders = HttpHeaders()
-        httpHeaders.contentType = MediaType.APPLICATION_PDF
-        return httpHeaders
+        return ResponseEntity(pdfContents, ControllerUtils.getHttpHeadersForPdf(), HttpStatus.OK)
     }
 }
