+++ src/main/proto/user_service.proto
@@ -59,11 +59,6 @@ message CoopResponse {
     string logo = 4;
 }
 
-message GetActiveUsersRequest {
-    string uuid = 1;
-    string coop = 2;
-}
-
 message UserExtendedResponse {
     string uuid = 1;
     string first_name = 2;
@@ -75,7 +70,6 @@ message UserExtendedResponse {
     string dateOfIssue = 8;
     string dateOfExpiry = 9;
     string personalNumber = 10;
-    Role role = 11;
 }
 
 message UsersExtendedResponse {
@@ -90,5 +84,5 @@ service UserService {
     rpc GetTokenIssuers (CoopRequest) returns (UsersResponse);
     rpc GetUserWithInfo (GetUserRequest) returns (UserWithInfoResponse);
     rpc GetUsersByEmail (GetUsersByEmailRequest) returns (UsersResponse);
-    rpc GetAllActiveUsers(GetActiveUsersRequest) returns (UsersExtendedResponse);
+    rpc GetAllActiveUsers (CoopRequest) returns (UsersExtendedResponse);
 }
